backend/formlyze/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==5.1.7
django-allauth==0.61.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
drf-yasg==1.21.10
et_xmlfile==2.0.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gspread==6.2.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==78.1.0
simplejson==3.20.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
